Phase 0: audio foundation fixes
Extract pure lib/mime.ts (EXT_MIME + derived AUDIO_EXTENSIONS + extOf);
content.ts imports contentType from it. Add webm (audio/webm) both sides
(api EXT_MIME, web AUDIO_EXTENSIONS). CSP _headers gains media-src 'self'
blob: <content-origin> for cross-origin audio + local recording preview.

Tests: W0-1/W0-2 webm, W0-3 extension-authoritative pin, W0-4 full MIME
table relocated to mime.test.ts, G-1 CSP structural gate (csp.test.ts).
