+++ packages/web/public/_headers
@@ -3,4 +3,4 @@
   X-Content-Type-Options: nosniff
   X-Frame-Options: DENY
   Referrer-Policy: strict-origin-when-cross-origin
-  Content-Security-Policy: default-src 'self'; script-src 'self' 'sha256-GHA4sP6dQXpE9w+vcP6hL8UmlMqDJF0nmaRuONkKD8U='; style-src 'self' 'unsafe-inline' https://fonts.googleapis.com; img-src 'self' data:; font-src 'self' https://fonts.gstatic.com; connect-src 'self'; frame-src 'self' https://glance-content.YOUR-SUBDOMAIN.workers.dev; object-src 'none'; base-uri 'self'; form-action 'self'
+  Content-Security-Policy: default-src 'self'; script-src 'self' 'sha256-GHA4sP6dQXpE9w+vcP6hL8UmlMqDJF0nmaRuONkKD8U='; style-src 'self' 'unsafe-inline' https://fonts.googleapis.com; img-src 'self' data:; media-src 'self' blob: https://glance-content.YOUR-SUBDOMAIN.workers.dev; font-src 'self' https://fonts.gstatic.com; connect-src 'self'; frame-src 'self' https://glance-content.YOUR-SUBDOMAIN.workers.dev; object-src 'none'; base-uri 'self'; form-action 'self'
