+++ packages/web/src/lib/audio.ts
@@ -2,7 +2,7 @@
 
 // Mirrors the audio EXT_MIME entries the content worker serves (packages/api/src/content.ts) —
 // same extensions, same authority rule (extension decides, not the stored/uploaded MIME).
-const AUDIO_EXTENSIONS = new Set(['mp3', 'wav', 'm4a', 'ogg', 'oga', 'flac', 'aac'])
+const AUDIO_EXTENSIONS = new Set(['mp3', 'wav', 'm4a', 'ogg', 'oga', 'flac', 'aac', 'webm'])
 
 /** True when `path`'s extension is one the content worker serves as audio. Used by the viewer
  *  to pick the AudioView (native `<audio>` player) over the sandboxed HTML iframe. */
